requirements.txt
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
streamlit>=1.42.0
plotly>=5.15.0
pandas>=2.1.0
supabase>=2.0.0
rich>=14.0.0
